Start refactoring lisk-codec protocol-specs
