+++ protocol-specs/generators/lisk_codec/types_generators/index.js
@@ -1,0 +1,28 @@
+/*
+ * Copyright © 2020 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ */
+const generateValidBooleanEncodings = require('./booleans');
+const generateValidNumberEncodings = require('./numbers');
+const generateValidStringEncodings = require('./strings');
+const generateValidBytesEncodings = require('./bytes');
+const generateValidObjectEncodings = require('./objects');
+const generateValidArrayEncodings = require('./arrays');
+
+module.exports = {
+	generateValidBooleanEncodings,
+	generateValidNumberEncodings,
+	generateValidStringEncodings,
+	generateValidBytesEncodings,
+	generateValidObjectEncodings,
+	generateValidArrayEncodings,
+};
